py-monbot/requirements.txt
discord.py == 1.3.1
pyyaml
pytz
pymongo[srv]